+++ app/sprinkles/account/src/Model/Permission.php
@@ -3,14 +3,13 @@
  * UserFrosting (http://www.userfrosting.com)
  *
  * @link      https://github.com/userfrosting/UserFrosting
- * @copyright Copyright (c) 2013-2016 Alexander Weissman
+ * @copyright Copyright (c) 2013-2017 Alexander Weissman
  * @license   https://github.com/userfrosting/UserFrosting/blob/master/licenses/UserFrosting.md (MIT License)
  */
 namespace UserFrosting\Sprinkle\Account\Model;
 
 use Illuminate\Database\Capsule\Manager as Capsule;
 use UserFrosting\Sprinkle\Core\Model\UFModel;
-use UserFrosting\Sprinkle\Core\Model\Relations\BelongsToManyThrough;
 
 /**
  * Permission Class.
@@ -100,30 +99,20 @@ class Permission extends UFModel
 
     /**
      * Get a list of users who have this permission, along with a list of roles through which each user has the permission.
+     *
+     * @return \UserFrosting\Sprinkle\Core\Model\Relations\BelongsToManyThrough
      */
     public function users()
     {
         /** @var UserFrosting\Sprinkle\Core\Util\ClassMapper $classMapper */
         $classMapper = static::$ci->classMapper;
 
-        // This relationship maps the permission to its roles.
-        $intermediateRelationship = $this->belongsToMany($classMapper->getClassMapping('role'), 'permission_roles', 'permission_id', 'role_id', 'roles')
-            ->withPivot('permission_id');
-
-        // Next, we need to find the users who have each of these roles.
-        
-        // If no relationship name was passed, we will pull backtraces to get the
-        // name of the calling function. We will use that function name as the
-        // title of this relation since that is a great convention to apply.
-        $relation = $this->guessBelongsToManyRelation();
-
-        // Now pass this query in to create a new `BelongsToManyThrough` relationship on this parent permission
-        $instance = $this->newRelatedInstance($classMapper->getClassMapping('user'));
-
-        $query = new BelongsToManyThrough(
-            $instance->newQuery(), $this, $intermediateRelationship, 'role_users', 'role_id', 'user_id', $relation
-        );
-
-        return $query;
+        return $this->belongsToManyThrough(
+            $classMapper->getClassMapping('user'),
+            $classMapper->getClassMapping('role'),
+            'permission_roles',
+            null,
+            null,
+            'role_users');
     }
 }