+++ app/sprinkles/account/src/Model/User.php
@@ -3,7 +3,7 @@
  * UserFrosting (http://www.userfrosting.com)
  *
  * @link      https://github.com/userfrosting/UserFrosting
- * @copyright Copyright (c) 2013-2016 Alexander Weissman
+ * @copyright Copyright (c) 2013-2017 Alexander Weissman
  * @license   https://github.com/userfrosting/UserFrosting/blob/master/licenses/UserFrosting.md (MIT License)
  */
 namespace UserFrosting\Sprinkle\Account\Model;
@@ -125,6 +125,8 @@ class User extends UFModel
 
     /**
      * Get all activities for this user.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\HasMany
      */
     public function activities()
     {
@@ -243,6 +245,8 @@ class User extends UFModel
 
     /**
      * Return this user's group.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
      */
     public function group()
     {
@@ -254,14 +258,15 @@ class User extends UFModel
 
     /**
      * Get the most recent activity for this user, based on the user's last_activity_id.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
      */
     public function lastActivity()
     {
         /** @var UserFrosting\Sprinkle\Core\Util\ClassMapper $classMapper */
         $classMapper = static::$ci->classMapper;
 
-        $query = $this->belongsTo($classMapper->getClassMapping('activity'), 'last_activity_id');
-        return $query;
+        return $this->belongsTo($classMapper->getClassMapping('activity'), 'last_activity_id');
     }
 
     /**
@@ -350,6 +355,8 @@ class User extends UFModel
 
     /**
      * Get all password reset requests for this user.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\HasMany
      */
     public function passwordResets()
     {
@@ -362,34 +369,26 @@ class User extends UFModel
     /**
      * Get all of the permissions this user has, via its roles.
      *
-     * @param string|null $slug If specified, filters by a specific slug.
-     * @todo Turn this into a full-fledged custom relation?
+     * @return \UserFrosting\Sprinkle\Core\Model\Relations\BelongsToManyThrough
      */
-    public function permissions($slug = null)
+    public function permissions()
     {
-        $result = Capsule::table('permissions')
-            ->select(
-                'permissions.id as id',
-                'roles.id as role_id',
-                'permissions.slug as slug',
-                'permissions.name as name',
-                'conditions',
-                'permissions.description as description')
-            ->join('permission_roles', 'permissions.id', '=', 'permission_roles.permission_id')
-            ->join('roles', 'permission_roles.role_id', '=', 'roles.id')
-            ->join('role_users', 'role_users.role_id', '=', 'roles.id')
-            ->where('role_users.user_id', '=', $this->id);
+        /** @var UserFrosting\Sprinkle\Core\Util\ClassMapper $classMapper */
+        $classMapper = static::$ci->classMapper;
 
-        if ($slug) {
-            $result = $result->where('permissions.slug', $slug);
-        }
-
-        return $result;
+        return $this->belongsToManyThrough(
+            $classMapper->getClassMapping('permission'),
+            $classMapper->getClassMapping('role'),
+            'role_users',
+            null,
+            null,
+            'permission_roles');
     }
 
     /**
      * Get all roles to which this user belongs.
      *
+     * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
      */
     public function roles()
     {
@@ -399,6 +398,11 @@ class User extends UFModel
         return $this->belongsToMany($classMapper->getClassMapping('role'), 'role_users', 'user_id', 'role_id')->withTimestamps();
     }
 
+    /**
+     * Get this user's roles, but only those that have a particular permission (specified elsewhere in the query).
+     *
+     * @return \UserFrosting\Sprinkle\Core\Model\Relations\BelongsToManyConstrained
+     */
     public function rolesWithPermission()
     {
         /** @var UserFrosting\Sprinkle\Core\Util\ClassMapper $classMapper */
@@ -430,6 +434,9 @@ class User extends UFModel
 
     /**
      * Joins the user's most recent activity directly, so we can do things like sort, search, paginate, etc.
+     *
+     * @param \Illuminate\Database\Eloquent\Builder $query
+     * @return \Illuminate\Database\Eloquent\Builder
      */
     public function scopeJoinLastActivity($query)
     {